Resilience Factor and Extended RF built
I am having a philosophical crisis. The QR metric finds the worst spot and anchors the times down. The RF metric uses the times as defined. I think we need the metrics to pick the spots right? That's what you would do if you were handed a performance plot. All you would know is that the performance went south and then recovered.